+++ src/marine_qc/qc_grouped_reports.py
@@ -7,8 +7,6 @@ Module containing QC functions for quality control of grouped marine reports.
 from __future__ import annotations
 import itertools
 import math
-from collections.abc import Sequence
-from typing import cast
 
 import numpy as np
 import pandas as pd
@@ -18,6 +16,7 @@ from .auxiliary import (
     SequenceFloatType,
     SequenceIntType,
     convert_units,
+    ensure_arrays,
     failed,
     inspect_arrays,
     isvalid,
@@ -104,12 +103,12 @@ class SuperObsGrid:
         self.nobs = np.zeros((360, 180, 73))  # type: np.ndarray
 
     @convert_date(["month", "day"])
-    @inspect_arrays(["lats", "lons", "values", "month", "day"])
+    @inspect_arrays(["lat", "lon", "value", "month", "day"])
     def add_multiple_observations(
         self,
-        lats: SequenceFloatType,
-        lons: SequenceFloatType,
-        values: SequenceFloatType,
+        lat: SequenceFloatType,
+        lon: SequenceFloatType,
+        value: SequenceFloatType,
         date: SequenceDatetimeType | None = None,
         month: SequenceFloatType | None = None,
         day: SequenceFloatType | None = None,
@@ -119,11 +118,11 @@ class SuperObsGrid:
 
         Parameters
         ----------
-        lats : array-like of float, shape (n,)
+        lat : array-like of float, shape (n,)
             1-dimensional latitude array.
-        lons : array-like of float, shape (n,)
+        lon : array-like of float, shape (n,)
             1-dimensional longitude array.
-        values : array-like of float, shape (n,)
+        value : array-like of float, shape (n,)
             1-dimensional anomaly array.
         date : array-like of datetime, shape (n,), optional
             1-dimensional datetime array.
@@ -136,39 +135,24 @@ class SuperObsGrid:
 
         Raises
         ------
-        ValueError
+        TypeError
             If `inspect_arrays` does not return np.ndarrays.
 
         Notes
         -----
         The observations should be anomalies.
         """
-        if not isinstance(lats, np.ndarray):
-            raise TypeError(f"'lats' must be a numpy.ndarray, got {type(lats).__name__}")
-        if not isinstance(lons, np.ndarray):
-            raise TypeError(f"'lons' must be a numpy.ndarray, got {type(lons).__name__}")
-        if not isinstance(values, np.ndarray):
-            raise TypeError(f"'values' must be a numpy.ndarray, got {type(values).__name__}")
-        if not isinstance(month, np.ndarray):
-            raise TypeError(f"'month' must be a numpy.ndarray, got {type(month).__name__}")
-        if not isinstance(day, np.ndarray):
-            raise TypeError(f"'day' must be a numpy.ndarray, got {type(day).__name__}")
-
-        value_arr: np.ndarray = np.atleast_1d(values).astype(float)
-        lat_arr: np.ndarray = np.atleast_1d(lats).astype(float)
-        lon_arr: np.ndarray = np.atleast_1d(lons).astype(float)
-        month_arr: np.ndarray = np.atleast_1d(month if month is not None else -1).astype(float)
-        day_arr: np.ndarray = np.atleast_1d(day if day is not None else -1).astype(float)
+        lat_arr, lon_arr, value_arr, month_arr, day_arr = ensure_arrays(lat=lat, lon=lon, value=value, month=month, day=day)
 
         month_arr = np.where(np.isnan(month_arr), -1, month_arr).astype(int)
         day_arr = np.where(np.isnan(day_arr), -1, day_arr).astype(int)
 
         valid: np.ndarray = (
-            np.atleast_1d(isvalid(lats))
-            & np.atleast_1d(isvalid(lons))
-            & np.atleast_1d(isvalid(month))
-            & np.atleast_1d(isvalid(day))
-            & np.atleast_1d(isvalid(values))
+            np.atleast_1d(isvalid(lat_arr))
+            & np.atleast_1d(isvalid(lon_arr))
+            & np.atleast_1d(isvalid(month_arr))
+            & np.atleast_1d(isvalid(day_arr))
+            & np.atleast_1d(isvalid(value_arr))
         )
         valid &= (month_arr >= 1) & (month_arr <= 12)
 
@@ -501,7 +485,7 @@ class SuperObsGrid:
 
 
 @post_format_return_type(["value"])
-@inspect_arrays(["lat", "lon", "date", "value"])
+@inspect_arrays(["lat", "lon", "date", "value", "climatology"])
 @convert_units(lat="degrees", lon="degrees")
 @inspect_climatology("climatology")
 def do_mds_buddy_check(
@@ -559,6 +543,11 @@ def do_mds_buddy_check(
         1-dimensional array containing QC flags.
         1 if buddy check fails, 0 otherwise.
 
+    Raises
+    ------
+    TypeError
+        If `inspect_arrays` does not return np.ndarrays.
+
     Notes
     -----
     The limits, number_of_obs_thresholds, and multipliers parameters are rather complex. The buddy check basically
@@ -579,13 +568,6 @@ def do_mds_buddy_check(
     * number_of_obs_thresholds = [[0, 5, 15, 100], [0], [0, 5, 15, 100], [0]]
     * multipliers = [[4.0, 3.5, 3.0, 2.5], [4.0], [4.0, 3.5, 3.0, 2.5], [4.0]]
     """
-    lat = cast(np.ndarray, lat)
-    lon = cast(np.ndarray, lon)
-    value = cast(np.ndarray, value)
-    climatology = cast(np.ndarray, climatology)
-
-    anoms = value - climatology
-
     if len(limits) != len(number_of_obs_thresholds) and len(limits) != len(multipliers):
         raise ValueError("Input parameter lists are not equal length")
 
@@ -593,38 +575,42 @@ def do_mds_buddy_check(
         if len(thresholds) != len(multipliers[i]):
             raise ValueError("Number of obs thresholds and multipliers have different shapes")
 
+    lat_arr, lon_arr, date_arr, value_arr, climatology_arr = ensure_arrays(lat=lat, lon=lon, date=date, value=value, climatology=climatology)
+
+    anoms_arr = value_arr - climatology_arr
+
     # calculate superob averages and numbers of observations
     grid = SuperObsGrid()
-    grid.add_multiple_observations(lat, lon, anoms, date=date)
+    grid.add_multiple_observations(lat=lat_arr, lon=lon_arr, value=anoms_arr, date=date_arr)
     grid.get_buddy_limits_with_parameters(standard_deviation, limits, number_of_obs_thresholds, multipliers)
 
-    numobs = len(lat)
+    numobs = len(lat_arr)
     qc_outcomes = np.zeros(numobs) + untested
 
     if ignore_indexes is None:
         ignore_indexes = []
 
-    valid_mask: np.ndarray = np.atleast_1d(isvalid(anoms))
+    valid_mask: np.ndarray = np.atleast_1d(isvalid(anoms_arr))
 
     # finally loop over all reports and update buddy QC
     for i in range(numobs):
         if i in ignore_indexes or not valid_mask[i]:
             continue
 
-        lat_: float = lat[i]
-        lon_: float = lon[i]
-        mon: pd.Timestamp = pd.Timestamp(date[i]).month
-        day: pd.Timestamp = pd.Timestamp(date[i]).day
+        lat_i: float = lat_arr[i]
+        lon_i: float = lon_arr[i]
+        mon_i: pd.Timestamp = pd.Timestamp(date_arr[i]).month
+        day_i: pd.Timestamp = pd.Timestamp(date_arr[i]).day
 
         # if the SST anomaly differs from the neighbour average by more than the calculated range then reject
-        x = anoms[i]
-        bm = grid.get_buddy_mean(lat_, lon_, mon, day)
-        bsd = grid.get_buddy_stdev(lat_, lon_, mon, day)
+        anom_i = anoms_arr[i]
+        bm = grid.get_buddy_mean(lat_i, lon_i, mon_i, day_i)
+        bsd = grid.get_buddy_stdev(lat_i, lon_i, mon_i, day_i)
 
         qc_outcomes[i] = passed
         if bsd == 500.0 or np.isnan(bsd):
             qc_outcomes[i] = untestable
-        elif abs(x - bm) >= bsd:
+        elif abs(anom_i - bm) >= bsd:
             qc_outcomes[i] = failed
 
     del grid
@@ -633,7 +619,7 @@ def do_mds_buddy_check(
 
 
 @post_format_return_type(["value"])
-@inspect_arrays(["lat", "lon", "date", "value"])
+@inspect_arrays(["lat", "lon", "date", "value", "climatology"])
 @convert_units(lat="degrees", lon="degrees")
 @inspect_climatology("climatology")
 def do_bayesian_buddy_check(
@@ -653,7 +639,7 @@ def do_bayesian_buddy_check(
     maximum_anomaly: float,
     fail_probability: float,
     ignore_indexes: list[int] | None = None,
-) -> Sequence[int]:
+) -> SequenceIntType:
     """
     Do the Bayesian buddy check.
 
@@ -708,6 +694,11 @@ def do_bayesian_buddy_check(
         1-dimensional array containing passed, failed or untestable flags. Untestable flags will be set if there
         are no buddies in the specified limits.
 
+    Raises
+    ------
+    TypeError
+        If `inspect_arrays` does not return np.ndarrays.
+
     Notes
     -----
     In previous versions the default values for the parameters were:
@@ -720,13 +711,9 @@ def do_bayesian_buddy_check(
     * maximum_anomaly = 8.0
     * fail_probability = 0.3
     """
-    lat = cast(np.ndarray, lat)
-    lon = cast(np.ndarray, lon)
-    date = cast(np.ndarray, date)
-    value = cast(np.ndarray, value)
-    climatology = cast(np.ndarray, climatology)
+    lat_arr, lon_arr, date_arr, value_arr, climatology_arr = ensure_arrays(lat=lat, lon=lon, date=date, value=value, climatology=climatology)
 
-    numobs = len(lat)
+    numobs = len(lat_arr)
     p0 = prior_probability_of_gross_error
     q = quantization_interval
     sigma_m = one_sigma_measurement_uncertainty
@@ -740,12 +727,12 @@ def do_bayesian_buddy_check(
 
     # Return untestable if any parameters is invalid
     if p0 < 0.0 or p0 > 1.0 or q <= 0.0 or r_hi < r_lo or sigma_m < 0.0:
-        return cast(Sequence[int], np.zeros(numobs) + untestable)
+        return np.zeros(numobs) + untestable
 
-    anoms = value - climatology
+    anoms_arr = value_arr - climatology_arr
 
     grid = SuperObsGrid()
-    grid.add_multiple_observations(lat, lon, anoms, date=date)
+    grid.add_multiple_observations(lat=lat_arr, lon=lon_arr, value=anoms_arr, date=date_arr)
     grid.get_new_buddy_limits(stdev1, stdev2, stdev3, limits, sigma_m, noise_scaling)
 
     qc_outcomes = np.zeros(numobs) + untested
@@ -757,16 +744,18 @@ def do_bayesian_buddy_check(
         if i in ignore_indexes:
             continue
 
-        if not isvalid(anoms[i]):
+        anom_i = anoms_arr[i]
+
+        if not isvalid(anom_i):
             continue
 
-        lat_ = lat[i]
-        lon_ = lon[i]
-        mon = pd.Timestamp(date[i]).month
-        day = pd.Timestamp(date[i]).day
+        lat_i = lat_arr[i]
+        lon_i = lon_arr[i]
+        mon_i = pd.Timestamp(date_arr[i]).month
+        day_i = pd.Timestamp(date_arr[i]).day
 
         # Calculate the probability of gross error given the set-up
-        buddy_stdev = grid.get_buddy_stdev(lat_, lon_, mon, day)
+        buddy_stdev = grid.get_buddy_stdev(lat_i, lon_i, mon_i, day_i)
 
         try:
             ppp = p_gross(
@@ -774,8 +763,8 @@ def do_bayesian_buddy_check(
                 q,
                 r_hi,
                 r_lo,
-                anoms[i],
-                grid.get_buddy_mean(lat_, lon_, mon, day),
+                anom_i,
+                grid.get_buddy_mean(lat_i, lon_i, mon_i, day_i),
                 buddy_stdev,
             )
         except (ValueError, ZeroDivisionError):
@@ -792,4 +781,4 @@ def do_bayesian_buddy_check(
 
     del grid
 
-    return cast(Sequence[int], qc_outcomes)
+    return qc_outcomes